Increase epsilon for more flexible stop vs waypoint connections
Currently, the loop continues until it finds the waypoint which is
actually the next stop. However, the previous epsilon was so strict and
the ISC/pennrides data is so messy that it would not find it sometimes.
Increasing the epsilon to 0.0002 fixes the bug, and since stops are
never very close to each other this should not be a problem.