Trim whitespace from room IDs in Laundry
The Android app is sending laundry room preferences data to the server as an ArrayList, but the ArrayList toString() method inserts ", " between each element of the ArrayList. As a result, the server is only able to parse the first element and all subsequent elements fail to be parsed because of the whitespace. This should fix the issue.